完成添加订单功能 和展示订单接口 但是前端没有调用  暂时停止该项目

+++ src/main/java/com/optimus/dto/results/OrderDetailResultsDTO.java
@@ -10,7 +10,7 @@ public class OrderDetailResultsDTO implements Serializable {
   private static final long serialVersionUID = 5492390941698795667L;
   private Long id;
   private Integer orderId;
-  private Integer orderNumber;
+  private Long orderNumber;
   private String storeName;
   private Integer productNumber;
   private Double productPrice;
